hit a bit of a roadblock, losing hope
Not sure if it's worth continuing this ambitious branch. FOLFormula
requires its subformulas to be FOLFormulas, of course, but dLFormula
will allow general dLFormulas, which means that the infrastructure for
FOLFormula will have to be replicated by dLFormula. This pretty much
defeats the entire point of this ambitious branch.

Instead, I will go and make my original dL as clean and organized as
possible.