Linearity checking and coefficient extraction works now.
The next step is to generate queries for matlab to find
linear Barrier/Lyapunov functions from the linear coefficients.

I would also like to:
-> Start using jUnit to incorporate more thorough testing

I backed down on making the ArithmeticAnalyzer into a decorator
for its corresponding structure. I couldn't convince myself that
it was a good idea. I also couldn't convince myself to put that
functionality into the different term types, but maybe it is a good
idea. I just sort of get trapped in this oscillatory mental state
where I can't decide which method is better--so for now I left it
as an external analytics library, though that is not good
object-oriented design.