Fixes to getBoundVariables and MathematicaInterface
getBoundVariables was losing the bound variables of subformulas in the
quantified variable case. MathematicaInterface was modified so that the
instance found by findInstance is in decimals, so that I don't have to
parse the horribleness of square root and all that. By the way,
checkValidity will be hard to use, unless I add something... like
InputForm?